Refactor GUI module to remove unused code and imports
